Trust the Feature test macro.
To determine if a feature in the library exists, we used to do the following:

#if defined (feature_test_macro) || CPPxx_OR_LATER // xx is 14/17/20

However, most compilers do not fully implement the feature of standard library for some time after the __cplusplus version bump up. This caused a compile error saying that it was using a non-existent std symbol.

After this commit, we will trust the feature test macro. Therefore, when the macro does not exist, it is judged that there is no feature in the standard library.

For gcc and clang, there's probably no problem. Since feature test macro is implemented in visual Studio 2017 or later in msvc, it is treated as if the C ++ 14 library function does not exist in 2015.